requirements.txt
alembic==1.11.3
amqp==5.1.1
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-default-group-wheel==1.2.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pytest==7.4.0
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlite-fts4==1.0.3
sqlite-utils==3.35
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
